refactor: fix conflicts during merge
