+++ TexTris/src/main/java/org/Stech/SE5/Model/data/config.json
@@ -1,15 +1,1 @@
-{
-  "GameMode": "BASIC",
-  "GameDifficulty": "NORMAL",
-  "BoardSize": "MEDIUM",
-  "BoardWidth": 10,
-  "BoardHeight": 20,
-  "GameSpeed": 1.0,
-  "ColorBlindMode": false,
-  "KeyBindingLength": 7,
-  "moveBlockLeft": 65,
-  "moveBlockRight": 68,
-  "moveBlockDown": 83,
-  "clockwiseRotate": 87,
-  "dropBlockAtOnce": 32
-}
+{"GameMode":"BASIC","GameDifficulty":"NORMAL","BoardSize":"MEDIUM","BoardWidth":10,"BoardHeight":20,"GameSpeed":1.0,"ColorBlindMode":false,"KeyBindingLength":7,"moveBlockLeft":37,"moveBlockRight":39,"moveBlockDown":40,"clockwiseRotate":38,"dropBlockAtOnce":32}
